Major Updates and Added References
Added the bibliography for the paper (sajanptl_shurjo.bib) and continued
making edits to the paper. Rewrote the problem and innovation sections
of the SLAM++ paper as well as added references appropriately throughout
the paper. Cleaned up the Latex file for the paper as well. Edited the
conclusions for GF-SLAM. Added overall paper conclusion as well.